added warning for --mmseqs-max-seqs saturation

+++ nail/src/io/fasta.rs
@@ -170,7 +170,7 @@ impl Fasta {
         self.buffer.resize(offset.n_bytes, 0u8);
 
         self.file
-            .seek(SeekFrom::Start(offset.start as u64))
+            .seek(SeekFrom::Start(offset.start))
             .expect("failed to seek in Fasta::get()");
 
         self.file
